Re-added GZIP classifier
It isn't really easy to determine the compressed size of a gzip, the
only way I'm aware of is to fully decompress and keep track in the
decompressor. We don't have this luxury and gzip doesn't emit that data
after successful decompression. Therefore we changed the code that for
some blobs it might be possible to extract, but not to carve the files
(i.e., "None" is permissible as a filesize).