I hate BZIP2
The format is horrifically documented and bzcat apparently seeks to the
beginning of the file when we pass a file descriptor to the process, so
we need to pass data between the processes (reading from the file,
passing into stdin of the new process). I've given up on determining
compressed length for bzip2 altogether, the format is INSANELY
complicated even without the bit stuffing.